+++ apps/otoge-bookmark-api/src/games/games.controller.ts
@@ -18,7 +18,7 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { GameEntity } from './entities/game.entity';
-import { JwtAuthGuard } from '@/auth/jwt-auth.guard';
+import { JwtAuthGuard } from '@/common/guards/jwt-auth.guard';
 
 @Controller('games')
 @UseGuards(JwtAuthGuard)
