+++ apps/otoge-bookmark-api/src/songs/songs.controller.ts
@@ -18,7 +18,7 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { SongEntity } from './entities/song.entity';
-import { JwtAuthGuard } from '@/auth/jwt-auth.guard';
+import { JwtAuthGuard } from '@/common/guards/jwt-auth.guard';
 
 @Controller('songs')
 @UseGuards(JwtAuthGuard)
